Adds Documentation about app/ structure
Create a useful introduction for people to get up and running
in contributing adidtional modules to the site.

This could go in a seperate Wiki within GitHub but I think it make's more sense to keep it front and center below the codebase. Perhaps if this becomes more long winded in the future it can be moved.